fix(*): evaluation UX inconsistency
